Add org.utbot.spring.api
