+++ utbot-instrumentation/src/main/kotlin/org/utbot/instrumentation/process/InstrumentedProcessMain.kt
@@ -46,8 +46,9 @@ internal object HandlerClassesLoader : URLClassLoader(emptyArray()) {
      *  - we want org.utbot.spring to be loaded by [HandlerClassesLoader] so it can use Spring directly
      */
     override fun loadClass(name: String, resolve: Boolean): Class<*> {
-        // TODO figure out why system class loader can find org.utbot.spring
-        if (name.startsWith("org.slf4j") || name.startsWith("org.utbot.spring")) {
+        // TODO extract `utbot-spring-commons-api` into a separate module to not mess around with class loader
+        if (name.startsWith("org.slf4j") ||
+            (name.startsWith("org.utbot.spring") && !name.startsWith("org.utbot.spring.api"))) {
             return (findLoadedClass(name) ?: findClass(name)).apply {
                 if (resolve) resolveClass(this)
             }
