+++ utbot-spring-commons/src/main/kotlin/org/utbot/spring/api/repositoryWrapper/RepositoryInteraction.kt
@@ -1,0 +1,14 @@
+package org.utbot.spring.api.repositoryWrapper
+
+import java.lang.reflect.Method
+
+data class RepositoryInteraction(
+    val beanName: String,
+    val method: Method,
+    val args: List<Any?>,
+    val result: Result<Any?>
+) {
+    companion object {
+        val recordedInteractions = mutableListOf<RepositoryInteraction>()
+    }
+}
