+++ utbot-spring-commons/src/main/kotlin/org/utbot/spring/repositoryWrapper/RepositoryWrapperInvocationHandler.kt
@@ -1,5 +1,6 @@
 package org.utbot.spring.repositoryWrapper
 
+import org.utbot.spring.api.repositoryWrapper.RepositoryInteraction
 import java.lang.reflect.InvocationHandler
 import java.lang.reflect.Method
 
@@ -10,7 +11,7 @@ class RepositoryWrapperInvocationHandler(
     override fun invoke(proxy: Any, method: Method, args: Array<out Any?>?): Any? {
         val nonNullArgs = args ?: emptyArray()
         val result = runCatching { method.invoke(originalRepository, *nonNullArgs) }
-        RepositoryInteraction.repositoryInteractions.add(
+        RepositoryInteraction.recordedInteractions.add(
             RepositoryInteraction(beanName, method, nonNullArgs.toList(), result)
         )
         return result.getOrThrow()
